requirements.txt
torch
transformers
bitsandbytes
accelerate
openai
tqdm
sentence-transformers
faiss-gpu
scikit-learn
numpy
peft
rouge-score